refactor(api): split repositories.rb into per-resource files
Move each repository class into db/repositories/<resource>.rb and the
shared mixins into db/repositories/mixins.rb, matching the per-resource
file layout already used by the API and Service layers. repositories.rb
becomes an aggregator that requires them all, so `require "db/repositories"`
callers are unchanged.

Closes: #52
